Further improve implementation
